requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
feedparser
beautifulsoup4
requests
google-generativeai
schedule
lxml
